#14 - Removed Spring Boot DevTools.
Removed the Spring Boot Dev Tools as the classloader separation causes Salespoint repositories bootstrap to fail for ones in package scope. The JDK proxy mechanism enforces visibility between classloaders and rejects the creation of the proxy for a type invisible to the current one.

This has been reported to the team and has a solution pending.